+++ lib/graph_weaver/railtie.rb
@@ -24,6 +24,6 @@ class GraphWeaver::Railtie < Rails::Railtie
   end
 
   initializer "graph_weaver.load_generated", after: :load_config_initializers do
-    GraphWeaver.load_generated! if Dir.exist?(GraphWeaver.generated_path)
+    GraphWeaver.load_generated! if GraphWeaver.generated_paths.any? { |path| Dir.exist?(path) }
   end
 end